Change in UI Dialogs

+++ lib/ui/widget/delete-member-dialog.widget.dart
@@ -3,8 +3,10 @@ import 'package:daruma/redux/state.dart';
 import 'package:daruma/redux/index.dart';
 import 'package:daruma/services/bloc/member.bloc.dart';
 import 'package:daruma/services/networking/index.dart';
+import 'package:daruma/util/colors.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_redux/flutter_redux.dart';
+import 'package:sweet_alert_dialogs/sweet_alert_dialogs.dart';
 
 class DeleteMemberDialog extends StatelessWidget {
   final Member member;
@@ -35,46 +37,56 @@ class DeleteMemberDialog extends StatelessWidget {
           if (snapshot.hasData) {
             switch (snapshot.data.status) {
               case Status.LOADING:
-                return Center(child: CircularProgressIndicator());
+              return RichAlertDialog(
+                  alertTitle: richTitle("Cargando"),
+                  alertSubtitle: richSubtitle("Se esta eliminado el miembro..."),
+                  alertType: RichAlertType.CUSTOM,
+                  dialogIcon: Icon(Icons.access_time, color: redPrimaryColor,),
+                );
                 break;
 
               case Status.COMPLETED:
-                return Container(
-                  height: 300.0, // Change as per your requirement
-                  width: 300.0,
-                  child: Row(
-                    children: <Widget>[
-                      Text("Delete completed!"),
-                      FlatButton(
-                        onPressed: () {
-                          vm.deleteMember(this.member);
+              return RichAlertDialog(
+                  alertTitle: richTitle("¡Completado!"),
+                  alertSubtitle: richSubtitle("Miembro eliminado correctamente"),
+                  alertType: RichAlertType.SUCCESS,
+                  actions: <Widget>[
+                    FlatButton(
+                      child: Text("OK"),
+                      onPressed: () {
+                        vm.deleteMember(this.member);
                           Navigator.of(context).pop();
-                        },
-                        child: Text(
-                          "Exit",
-                        ),
-                      )
-                    ],
-                  ),
+                      },
+                    )
+                  ],
                 );
                 break;
               case Status.ERROR:
-                return Container(
-                  height: 300.0, // Change as per your requirement
-                  width: 300.0,
-                  child: Row(
-                    children: <Widget>[
-                      Text("Delete ERROR!"),
-                      FlatButton(
-                        onPressed: () {
-                          Navigator.pop(context, true);
-                        },
-                        child: Text(
-                          "Exit",
-                        ),
-                      )
-                    ],
-                  ),
+              var errorMessage = snapshot.data.message;
+              var codeStatus = int.parse(errorMessage.substring(31, 34));
+
+              var errorSubtitle = "Se ha producido un error";
+
+              if(codeStatus == 403){
+                errorSubtitle = "No te puedes eliminar a ti mismo";
+              } else if(codeStatus == 404){
+                errorSubtitle = "No se puede eliminar ";
+              } else if(codeStatus == 400){
+                errorSubtitle = "El miembro está involucrado en un gasto";
+              }
+              
+              return RichAlertDialog(
+                  alertTitle: richTitle("Error"),
+                  alertSubtitle: richSubtitle(errorSubtitle),
+                  alertType: RichAlertType.ERROR,
+                  actions: <Widget>[
+                    FlatButton(
+                      child: Text("OK"),
+                      onPressed: () {
+                        Navigator.pop(context, true);
+                      },
+                    )
+                  ],
                 );
                 break;
             }
